# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Takas Backend API",
    "description": "İkinci el takas mobil uygulaması için REST API",
    "version": "1.0.0",
    "contact": {
      "name": "Takas Backend Team",
      "email": "info@takas.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Message": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "sender_id": {"type": "integer"},
          "receiver_id": {"type": "integer"},
          "trade_id": {"type": "integer"},
          "content": {"type": "string"},
          "is_read": {"type": "boolean"},
          "created_at": {"type": "string", "format": "date-time"}
        }
      }
    }
  },
  "paths": {
...
    "/messages": {
      "get": {
        "tags": ["Messages"],
        "summary": "Mesajları listele",
        "description": "Kullanıcının mesajlarını listeler",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer"}, "description": "Sayfa numarası"},
          {"name": "per_page", "in": "query", "schema": {"type": "integer"}, "description": "Sayfa başına mesaj sayısı"}
        ],
        "responses": {
          "200": {
            "description": "Mesaj listesi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "messages": {"type": "array", "items": {"$ref": "#/components/schemas/Message"}},
                    "total": {"type": "integer"},
                    "pages": {"type": "integer"},
                    "current_page": {"type": "integer"}
                  }
                }
              }
            }
          },
          "401": {"description": "Yetkilendirme gerekli"}
        }
      },
      "post": {
        "tags": ["Messages"],
        "summary": "Yeni mesaj gönder",
        "description": "Başka bir kullanıcıya mesaj gönderir",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["receiver_id", "content"],
                "properties": {
                  "receiver_id": {"type": "integer", "example": 2},
                  "trade_id": {"type": "integer", "example": 1},
                  "content": {"type": "string", "example": "Merhaba, ürün hakkında bilgi alabilir miyim?"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mesaj başarıyla gönderildi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string"},
                    "message_data": {"$ref": "#/components/schemas/Message"}
                  }
                }
              }
            }
          },
          "400": {"description": "Geçersiz veri"},
          "401": {"description": "Yetkilendirme gerekli"}
        }
      }
    }
  },
...
}